+++ spotify_playlists.txt.example
@@ -1,0 +1,2 @@
+https://open.spotify.com/playlist/37i9dQZF1DWWMOmoXKqHTD
+https://open.spotify.com/playlist/37i9dQZF1DWSqmBTGDYngZ
